BackEnd/lambda/requirements.txt
# AWS SDK for Python
boto3>=1.34.0
botocore>=1.34.0

# Environment variable management
python-dotenv>=1.0.0

# HTTP requests (for downloading transcripts)
requests>=2.31.0

# Google Gemini API (for audio processing lambda)
google-generativeai>=0.3.2
